pyproject.toml
[tool.poetry]
name = "instachecker"
version = "0.1.0"
description = ""
authors = ["Patrick Allen <prallen90@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
twilio = "^6.38.0"
python-dotenv = "^0.12.0"
pyyaml = "^5.3.1"
boto3 = "^1.12.39"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
pylint = "^2.4.4"
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
